fix: Restructured the next js routes
